Extract multimodal module into standalone llm-multimodal crate
- Create multimodal/ workspace crate with package name llm-multimodal
- Move vision processors, tracker, registry, and media handling
- Move tests and fixtures to multimodal/tests/
- Move generate_vision_golden.py script to multimodal/scripts/
- Update all path references in tests, scripts, and workflows
- Update .gitignore and .pre-commit-config.yaml
- Re-export llm_multimodal from src/multimodal/mod.rs for backwards compat
