+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["protocols", "reasoning_parser", "tool_parser", "workflow", "tokenizer", "auth", "mcp", "kv_index", "data_connector"]
+members = ["protocols", "reasoning_parser", "tool_parser", "workflow", "tokenizer", "auth", "mcp", "kv_index", "data_connector", "multimodal"]
 exclude = ["bindings/python"]
 resolver = "2"
 
@@ -137,6 +137,7 @@ smg-auth = { path = "auth", package = "auth" }
 smg-mcp = { path = "mcp", package = "mcp" }
 kv_index = { path = "kv_index" }
 smg-data-connector = { path = "data_connector", package = "data-connector" }
+llm-multimodal = { path = "multimodal" }
 
 # gRPC and Protobuf dependencies
 tonic = { version = "0.14.2", features = ["gzip", "transport"] }
