+++ multimodal/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "llm-multimodal"
+version = "0.1.0"
+edition = "2021"
+description = "Multimodal processing for vision and other modalities"
+license = "Apache-2.0"
+repository = "https://github.com/lightseekorg/smg"
+keywords = ["vision", "multimodal", "image-processing", "llm"]
+categories = ["multimedia::images", "science"]
+
+[lib]
+name = "llm_multimodal"
+
+[dependencies]
+async-trait.workspace = true
+base64 = "0.22"
+bytes = "1.8.0"
+image = { version = "0.25.4", default-features = false, features = ["png", "jpeg", "gif", "bmp", "ico", "tiff", "webp"] }
+ndarray = "0.16"
+once_cell = "1.21.3"
+reqwest = { workspace = true, features = ["stream"] }
+serde = { workspace = true, features = ["derive"] }
+serde_bytes = "0.11"
+serde_json.workspace = true
+thiserror.workspace = true
+tokio = { workspace = true, features = ["sync", "fs", "rt-multi-thread"] }
+tracing.workspace = true
+url = "2.5.4"
+
+# Workspace crates
+llm-tokenizer = { path = "../tokenizer", package = "tokenizer" }
+
+[dev-dependencies]
+npyz = { version = "0.8", features = ["npz"] }
+tempfile = "3.8"
+tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
