+++ multimodal/src/error.rs
