+++ multimodal/src/lib.rs
@@ -1,0 +1,22 @@
+pub mod error;
+pub mod media;
+pub mod registry;
+pub mod tracker;
+pub mod types;
+pub mod vision;
+
+pub use error::{MediaConnectorError, MultiModalError, MultiModalResult};
+pub use media::{ImageFetchConfig, MediaConnector, MediaConnectorConfig, MediaSource};
+pub use registry::{ModelProcessorSpec, ModelRegistry};
+pub use tracker::{AsyncMultiModalTracker, TrackerConfig, TrackerOutput};
+pub use types::{
+    ChatContentPart, ConversationSegment, ImageDetail, ImageFrame, ImageSize, ImageSource,
+    Modality, MultiModalData, MultiModalInputs, MultiModalTensor, MultiModalUUIDs, MultiModalValue,
+    PlaceholderHandle, PlaceholderMap, PlaceholderRange, PromptReplacement, TokenId, TrackedMedia,
+    DEFAULT_PLACEHOLDERS,
+};
+// Re-export vision processing components
+pub use vision::{
+    ImagePreProcessor, ImageProcessorRegistry, LlavaNextProcessor, LlavaProcessor,
+    PreProcessorConfig, PreprocessedImages, TransformError,
+};
