+++ multimodal/src/vision/mod.rs
