+++ multimodal/src/vision/preprocessor_config.rs
