+++ multimodal/src/vision/processors/mod.rs
