+++ multimodal/src/vision/processors/phi3_vision.rs
@@ -22,7 +22,7 @@
 use image::{imageops::FilterType, DynamicImage, GenericImageView, Rgb, RgbImage};
 use ndarray::{s, Array3, Array4, IxDyn};
 
-use crate::multimodal::vision::{
+use crate::vision::{
     image_processor::{ImagePreProcessor, PreprocessedImages},
     preprocessor_config::PreProcessorConfig,
     transforms::{self, TransformError},
@@ -339,7 +339,7 @@ impl ImagePreProcessor for Phi3VisionProcessor {
             .collect();
         model_specific.insert(
             "image_sizes".to_string(),
-            crate::multimodal::vision::image_processor::ModelSpecificValue::UintTensor {
+            crate::vision::image_processor::ModelSpecificValue::UintTensor {
                 data: image_sizes_data,
                 shape: vec![batch_size, 2],
             },
@@ -348,7 +348,7 @@ impl ImagePreProcessor for Phi3VisionProcessor {
         // num_img_tokens as list
         model_specific.insert(
             "num_img_tokens".to_string(),
-            crate::multimodal::vision::image_processor::ModelSpecificValue::UintVec(
+            crate::vision::image_processor::ModelSpecificValue::UintVec(
                 all_num_tokens.iter().map(|&t| t as u32).collect(),
             ),
         );
