+++ multimodal/src/vision/processors/phi4_vision.rs
@@ -38,7 +38,7 @@ use std::collections::HashSet;
 use image::{imageops::FilterType, DynamicImage, GenericImageView, Rgb, RgbImage};
 use ndarray::{s, Array2, Array3, Array4, IxDyn};
 
-use crate::multimodal::vision::{
+use crate::vision::{
     image_processor::{ImagePreProcessor, ModelSpecificValue, PreprocessedImages},
     preprocessor_config::PreProcessorConfig,
     transforms::{self, TransformError},
