+++ multimodal/tests/fixtures/images/grayscale.jpg
