+++ multimodal/tests/fixtures/images/large.jpg
