+++ multimodal/tests/fixtures/images/odd_dims.jpg
