+++ multimodal/tests/fixtures/images/small.jpg
