+++ multimodal/tests/fixtures/images/square.jpg
