+++ multimodal/tests/fixtures/images/tall.jpg
