+++ multimodal/tests/fixtures/images/tiny.jpg
