+++ multimodal/tests/fixtures/images/very_tall.jpg
