+++ multimodal/tests/fixtures/images/very_wide.jpg
