+++ multimodal/tests/fixtures/images/wide.jpg
