+++ multimodal/tests/multimodal_tracker_test.rs
@@ -1,11 +1,11 @@
 use std::{collections::HashMap, path::PathBuf, sync::Arc, time::Duration};
 
 use base64::{engine::general_purpose::STANDARD as BASE64_STANDARD, Engine};
-use reqwest::Client;
-use smg::multimodal::{
+use llm_multimodal::{
     AsyncMultiModalTracker, ChatContentPart, ConversationSegment, ImageFetchConfig, ImageSource,
     MediaConnector, MediaConnectorConfig, MediaSource, Modality, TrackerConfig,
 };
+use reqwest::Client;
 use tempfile::tempdir;
 
 const TINY_PNG_BASE64: &str =
