+++ multimodal/tests/vision_golden_tests.rs
@@ -11,17 +11,17 @@
 //!
 //! To regenerate golden outputs:
 //! ```bash
-//! python scripts/generate_vision_golden.py
+//! python multimodal/scripts/generate_vision_golden.py
 //! ```
 
 use std::{fs::File, io::Read, path::Path};
 
-use ndarray::{Array4, Array5};
-use smg::multimodal::vision::{
+use llm_multimodal::vision::{
     image_processor::ModelSpecificValue, ImagePreProcessor, Llama4VisionProcessor, LlavaProcessor,
     Phi3VisionProcessor, Phi4VisionProcessor, PixtralProcessor, PreProcessorConfig,
     Qwen2VLProcessor, Qwen3VLProcessor,
 };
+use ndarray::{Array4, Array5};
 
 /// Load a numpy .npz file and extract pixel_values
 fn load_golden_npz(path: &Path) -> Array4<f32> {
@@ -117,15 +117,16 @@ fn load_golden_num_tokens(path: &Path) -> usize {
 /// * `mode` - Either "llava" (standard CLIP) or "llava_pad" (expand-to-square mode)
 /// * `image_name` - Name of the test image (e.g., "square", "tall", "wide", "small")
 fn run_golden_test(mode: &str, image_name: &str) {
-    let golden_dir = Path::new("tests/fixtures/golden").join(mode);
-    let image_path = Path::new("tests/fixtures/images").join(format!("{}.jpg", image_name));
+    let golden_dir = Path::new("multimodal/tests/fixtures/golden").join(mode);
+    let image_path =
+        Path::new("multimodal/tests/fixtures/images").join(format!("{}.jpg", image_name));
 
     if !golden_dir.exists() || !image_path.exists() {
         eprintln!(
             "Golden test fixtures for {}/{} not found, skipping test",
             mode, image_name
         );
-        eprintln!("Run: python scripts/generate_vision_golden.py");
+        eprintln!("Run: python multimodal/scripts/generate_vision_golden.py");
         return;
     }
 
@@ -273,7 +274,7 @@ fn test_llava_pad_golden_grayscale() {
 
 #[test]
 fn test_llava_token_count() {
-    let golden_dir = Path::new("tests/fixtures/golden/llava");
+    let golden_dir = Path::new("multimodal/tests/fixtures/golden/llava");
 
     if !golden_dir.exists() {
         eprintln!("Golden test fixtures not found, skipping test");
@@ -320,15 +321,16 @@ fn load_golden_qwen2_vl_pixels(path: &Path) -> (Vec<f32>, (usize, usize)) {
 /// 2. num_tokens calculation is correct
 /// 3. Pixel values match after reshaping to patch format
 fn run_qwen2_vl_golden_test(image_name: &str) {
-    let golden_dir = Path::new("tests/fixtures/golden/qwen2_vl");
-    let image_path = Path::new("tests/fixtures/images").join(format!("{}.jpg", image_name));
+    let golden_dir = Path::new("multimodal/tests/fixtures/golden/qwen2_vl");
+    let image_path =
+        Path::new("multimodal/tests/fixtures/images").join(format!("{}.jpg", image_name));
 
     if !golden_dir.exists() || !image_path.exists() {
         eprintln!(
             "Golden test fixtures for qwen2_vl/{} not found, skipping test",
             image_name
         );
-        eprintln!("Run: python scripts/generate_vision_golden.py --model qwen2_vl");
+        eprintln!("Run: python multimodal/scripts/generate_vision_golden.py --model qwen2_vl");
         return;
     }
 
@@ -504,15 +506,16 @@ fn test_qwen2_vl_golden_grayscale() {
 /// - factor: 32 (vs 28)
 /// - normalization: [0.5, 0.5, 0.5] (vs CLIP)
 fn run_qwen3_vl_golden_test(image_name: &str) {
-    let golden_dir = Path::new("tests/fixtures/golden/qwen3_vl");
-    let image_path = Path::new("tests/fixtures/images").join(format!("{}.jpg", image_name));
+    let golden_dir = Path::new("multimodal/tests/fixtures/golden/qwen3_vl");
+    let image_path =
+        Path::new("multimodal/tests/fixtures/images").join(format!("{}.jpg", image_name));
 
     if !golden_dir.exists() || !image_path.exists() {
         eprintln!(
             "Golden test fixtures for qwen3_vl/{} not found, skipping test",
             image_name
         );
-        eprintln!("Run: python scripts/generate_vision_golden.py --model qwen3_vl");
+        eprintln!("Run: python multimodal/scripts/generate_vision_golden.py --model qwen3_vl");
         return;
     }
 
@@ -801,15 +804,16 @@ fn find_max_diff_location_5d(
 /// 3. num_img_tokens matches HuggingFace output
 /// 4. Pixel values match within tolerance
 fn run_phi3_vision_golden_test(image_name: &str) {
-    let golden_dir = Path::new("tests/fixtures/golden/phi3_vision");
-    let image_path = Path::new("tests/fixtures/images").join(format!("{}.jpg", image_name));
+    let golden_dir = Path::new("multimodal/tests/fixtures/golden/phi3_vision");
+    let image_path =
+        Path::new("multimodal/tests/fixtures/images").join(format!("{}.jpg", image_name));
 
     if !golden_dir.exists() || !image_path.exists() {
         eprintln!(
             "Golden test fixtures for phi3_vision/{} not found, skipping test",
             image_name
         );
-        eprintln!("Run: python scripts/generate_vision_golden.py --model phi3_vision");
+        eprintln!("Run: python multimodal/scripts/generate_vision_golden.py --model phi3_vision");
         return;
     }
 
@@ -1015,15 +1019,16 @@ fn load_phi4_image_sizes(path: &Path) -> Vec<(u32, u32)> {
 /// - Normalization: [0.5, 0.5, 0.5] (vs CLIP)
 /// - Default dynamic_hd: 36 (vs 16)
 fn run_phi4_vision_golden_test(image_name: &str) {
-    let golden_dir = Path::new("tests/fixtures/golden/phi4_vision");
-    let image_path = Path::new("tests/fixtures/images").join(format!("{}.jpg", image_name));
+    let golden_dir = Path::new("multimodal/tests/fixtures/golden/phi4_vision");
+    let image_path =
+        Path::new("multimodal/tests/fixtures/images").join(format!("{}.jpg", image_name));
 
     if !golden_dir.exists() || !image_path.exists() {
         eprintln!(
             "Golden test fixtures for phi4_vision/{} not found, skipping test",
             image_name
         );
-        eprintln!("Run: python scripts/generate_vision_golden.py --model phi4_vision");
+        eprintln!("Run: python multimodal/scripts/generate_vision_golden.py --model phi4_vision");
         return;
     }
 
@@ -1232,15 +1237,16 @@ fn load_llama4_pixels(path: &Path) -> (Vec<f32>, Vec<usize>) {
 /// - Normalization: [0.5, 0.5, 0.5] mean/std
 /// - Global tile added when num_tiles > 1
 fn run_llama4_vision_golden_test(image_name: &str) {
-    let golden_dir = Path::new("tests/fixtures/golden/llama4_vision");
-    let image_path = Path::new("tests/fixtures/images").join(format!("{}.jpg", image_name));
+    let golden_dir = Path::new("multimodal/tests/fixtures/golden/llama4_vision");
+    let image_path =
+        Path::new("multimodal/tests/fixtures/images").join(format!("{}.jpg", image_name));
 
     if !golden_dir.exists() || !image_path.exists() {
         eprintln!(
             "Golden test fixtures for llama4_vision/{} not found, skipping test",
             image_name
         );
-        eprintln!("Run: python scripts/generate_vision_golden.py --model llama4_vision");
+        eprintln!("Run: python multimodal/scripts/generate_vision_golden.py --model llama4_vision");
         return;
     }
 
@@ -1429,15 +1435,16 @@ fn load_pixtral_image_sizes(path: &Path) -> Vec<(usize, usize)> {
 /// - Normalization: CLIP mean/std
 /// - No tiling - single output per image
 fn run_pixtral_golden_test(image_name: &str) {
-    let golden_dir = Path::new("tests/fixtures/golden/pixtral");
-    let image_path = Path::new("tests/fixtures/images").join(format!("{}.jpg", image_name));
+    let golden_dir = Path::new("multimodal/tests/fixtures/golden/pixtral");
+    let image_path =
+        Path::new("multimodal/tests/fixtures/images").join(format!("{}.jpg", image_name));
 
     if !golden_dir.exists() || !image_path.exists() {
         eprintln!(
             "Golden test fixtures for pixtral/{} not found, skipping test",
             image_name
         );
-        eprintln!("Run: python scripts/generate_vision_golden.py --model pixtral");
+        eprintln!("Run: python multimodal/scripts/generate_vision_golden.py --model pixtral");
         return;
     }
