+++ src/multimodal/mod.rs
@@ -1,17 +1,6 @@
-pub mod error;
-pub mod media;
-pub mod registry;
-pub mod tracker;
-pub mod types;
-pub mod vision;
+//! Re-exports from the llm-multimodal crate.
+//!
+//! This module provides backwards-compatible access to multimodal functionality
+//! that has been extracted to the standalone `llm-multimodal` crate.
 
-pub use error::{MediaConnectorError, MultiModalError, MultiModalResult};
-pub use media::{ImageFetchConfig, MediaConnector, MediaConnectorConfig, MediaSource};
-pub use registry::{ModelProcessorSpec, ModelRegistry};
-pub use tracker::{AsyncMultiModalTracker, TrackerConfig, TrackerOutput};
-pub use types::*;
-// Re-export vision processing components
-pub use vision::{
-    ImagePreProcessor, ImageProcessorRegistry, LlavaNextProcessor, LlavaProcessor,
-    PreProcessorConfig, PreprocessedImages, TransformError,
-};
+pub use llm_multimodal::*;
