--- tests/multimodal/mod.rs
@@ -1,4 +1,0 @@
-//! Multimodal and vision integration tests
-
-mod multimodal_tracker_test;
-mod vision_golden_tests;
