--- tests/multimodal_tests.rs
@@ -1,6 +1,0 @@
-//! Multimodal integration tests
-
-#[path = "common/mod.rs"]
-pub mod common;
-
-mod multimodal;
